add JWT

+++ src/service/product.py
@@ -15,14 +15,31 @@ class Product(Common):
         database.main[collection].insert_one(product)
         return self.entity_response(product)
 
-    def put(self, product):
+    def put(self, product,current_user):
+        print(product['produtor_id'])
+        if product['produtor_id'] != current_user['id']:
+             return {
+                    'resultado': False,
+                    'mensagem': "erro ao atualizar produto,esse produto pertence a outro usuário"
+                  },403
+
         my_query = { '_id':  ObjectId(product['id']) }
         del product['id']
         new_values = { '$set': product}
         return database.main[collection].update_one(my_query, new_values) 
 
-    def delete(self, id):
+    def delete(self, id,current_user): 
+        product=database.main[collection].find_one({'_id':  ObjectId(id)})
+        if product['produtor_id'] != current_user['id']:
+           return {
+                    'resultado': False,
+                    'mensagem': "erro ao apagar produto,esse produto pertence a outro usuário"
+                  },403
         database.main[collection].delete_one({'_id':  ObjectId(id)})
+        return {
+                    'resultado': True,
+                    'mensagem': "Produto apagado com sucesso"
+                  }
     
     def get_one(self, id):
         product = database.main[collection].find_one({'_id':  ObjectId(id)})
